cmd/pebble: remove old compaction benchmarking tool
Remove the old compaction benchmarking tool. It was unused and will be
replaced as a part of #1865. Additionally, remove the FlushExternal
logic that existed solely to support the old compaction benchmarking
tool.
